+++ src/main/java/com/nextgis/maplib/map/MapEventSource.java
@@ -100,7 +100,7 @@ public class MapEventSource
 
     @Override
     public void onDrawFinished(
-            int id,
+            long id,
             float percent)
     {
         super.onDrawFinished(id, percent);
@@ -123,7 +123,7 @@ public class MapEventSource
         }
 
         Bundle bundle = new Bundle();
-        bundle.putInt(BUNDLE_ID_KEY, layer.getId());
+        bundle.putLong(BUNDLE_ID_KEY, layer.getId());
         bundle.putInt(BUNDLE_TYPE_KEY, EVENT_onLayerAdded);
 
         Message msg = new Message();
@@ -147,7 +147,7 @@ public class MapEventSource
         }
 
         Bundle bundle = new Bundle();
-        bundle.putInt(BUNDLE_ID_KEY, layer.getId());
+        bundle.putLong(BUNDLE_ID_KEY, layer.getId());
         bundle.putInt(BUNDLE_TYPE_KEY, EVENT_onLayerChanged);
 
         Message msg = new Message();
@@ -163,7 +163,7 @@ public class MapEventSource
      *         A deleted layer identificator
      */
     @Override
-    protected void onLayerDeleted(int id)
+    protected void onLayerDeleted(long id)
     {
         super.onLayerDeleted(id);
         if (mListeners == null) {
@@ -171,7 +171,7 @@ public class MapEventSource
         }
 
         Bundle bundle = new Bundle();
-        bundle.putInt(BUNDLE_ID_KEY, id);
+        bundle.putLong(BUNDLE_ID_KEY, id);
         bundle.putInt(BUNDLE_TYPE_KEY, EVENT_onLayerDeleted);
 
         Message msg = new Message();
@@ -234,7 +234,7 @@ public class MapEventSource
      * Send layers draw finished event to all listeners
      */
     protected void onLayerDrawFinished(
-            int id,
+            long id,
             float percent)
     {
         if (mListeners == null) {
@@ -243,7 +243,7 @@ public class MapEventSource
 
         Bundle bundle = new Bundle();
         bundle.putInt(BUNDLE_TYPE_KEY, EVENT_onLayerDrawFinished);
-        bundle.putInt(BUNDLE_ID_KEY, id);
+        bundle.putLong(BUNDLE_ID_KEY, id);
         bundle.putFloat(BUNDLE_DONE_KEY, percent);
 
         Message msg = new Message();
@@ -290,13 +290,13 @@ public class MapEventSource
                 for (MapEventListener listener : mListeners) {
                     switch (resultData.getInt(BUNDLE_TYPE_KEY)) {
                         case EVENT_onLayerAdded:
-                            listener.onLayerAdded(resultData.getInt(BUNDLE_ID_KEY));
+                            listener.onLayerAdded(resultData.getLong(BUNDLE_ID_KEY));
                             break;
                         case EVENT_onLayerDeleted:
-                            listener.onLayerDeleted(resultData.getInt(BUNDLE_ID_KEY));
+                            listener.onLayerDeleted(resultData.getLong(BUNDLE_ID_KEY));
                             break;
                         case EVENT_onLayerChanged:
-                            listener.onLayerChanged(resultData.getInt(BUNDLE_ID_KEY));
+                            listener.onLayerChanged(resultData.getLong(BUNDLE_ID_KEY));
                             break;
                         case EVENT_onExtentChanged:
                             listener.onExtentChanged(
@@ -306,7 +306,7 @@ public class MapEventSource
                             break;
                         case EVENT_onLayerDrawFinished:
                             listener.onLayerDrawFinished(
-                                    resultData.getInt(BUNDLE_ID_KEY),
+                                    resultData.getLong(BUNDLE_ID_KEY),
                                     resultData.getFloat(BUNDLE_DONE_KEY));
                             break;
                         case EVENT_onLayersReordered:
